% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate.R
\name{collate}
\alias{collate}
\title{Add columns to a data frame}
\usage{
collate(.data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Name-value pairs.
The name gives the name of the column in the output.

The value can be:
\itemize{
\item A vector of length 1, which will be recycled to the correct length.
\item A vector the same length as the current group (or the whole data frame
if ungrouped).
\item \code{NULL}, to remove the column.
\item A data frame or tibble, to create multiple columns in the output.
}}

\item{.before, .after}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, control where new columns
should appear (the default is to add to the right hand side). See
\code{\link[dplyr:relocate]{relocate()}} for more details.}
}
\value{
A data frame
}
\description{
A wrapper around \code{dplyr::mutate()} that restricts the scope to the creation of
new columns. It can be seen as a data masking variant of \code{dplyr::bind_cols()}.
}
\seealso{
populate
}
